Renamed test_config to config, since it's not actually testing the config file, it is the config for the tests.
