+++ tests/test_clinvar_to_evidence_strings.py
@@ -4,7 +4,7 @@ import unittest
 from eva_cttv_pipeline import consequence_type as CT
 from eva_cttv_pipeline import clinvar_to_evidence_strings
 from tests import test_clinvar
-import tests.test_config as test_config
+from tests import config
 
 
 def _get_mappings():
@@ -225,7 +225,7 @@ class TestGetConsequenceTypes(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.test_crm = test_clinvar.get_test_record().measures[0]
-        cls.consequence_type_dict = CT.process_consequence_type_file(test_config.snp_2_gene_file)
+        cls.consequence_type_dict = CT.process_consequence_type_file(config.snp_2_gene_file)
 
     def test_get_consequence_types(self):
         test_consequence_type = CT.ConsequenceType("ENSG00000163646", "stop_gained")
